# Faster GELU forward & backward using MUFU.TANH for SM7.5+