@ademeure 2024-07-31
Oops, /dev/cuda/ numbers are for non-MUFU.TANH part of the changes (e.g. faster way to do the derivative for backward) because I compiled without specifying the arch, so it's not SM7.5 😅 The val loss numbers for llm.c are correct though.

Can't test it until back on Tuesday but expecting it to be way faster