@ademeure 2024-08-07
I think there's an error for both /dev/cuda/common.h and test_gpt2.cu using an epsilon of 0.079 instead of 0.0079 for BF16 which makes the error threshold too high (missing some fairly large errors) - but even after fixing that, this seems to pass all the tests :)

The bandwidth calculations for gelu_backward were broken, and the batch size was wayyyy too small so it couldn't saturate H100 on either kernel, this is the correct performance (compiling for SM9.0):

```
FORWARD PASS - BEFORE
block_size   32 | time 0.3724 ms | bandwidth 1081.16 GB/s
block_size   64 | time 0.2472 ms | bandwidth 1628.73 GB/s
block_size  128 | time 0.2477 ms | bandwidth 1625.77 GB/s
block_size  256 | time 0.2496 ms | bandwidth 1613.01 GB/s
block_size  512 | time 0.2552 ms | bandwidth 1577.94 GB/s
block_size 1024 | time 0.2771 ms | bandwidth 1452.96 GB/s

FORWARD PASS - AFTER
block_size   32 | time 0.3738 ms | bandwidth 1077.24 GB/s
block_size   64 | time 0.2327 ms | bandwidth 1730.08 GB/s
block_size  128 | time 0.2327 ms | bandwidth 1730.10 GB/s
block_size  256 | time 0.2330 ms | bandwidth 1728.38 GB/s
block_size  512 | time 0.2340 ms | bandwidth 1721.00 GB/s
block_size 1024 | time 0.2370 ms | bandwidth 1699.07 GB/s

BACKWARD PASS - BEFORE
block_size   32 | time 0.4575 ms | bandwidth 1320.26 GB/s
block_size   64 | time 0.4231 ms | bandwidth 1427.44 GB/s
block_size  128 | time 0.4232 ms | bandwidth 1427.10 GB/s
block_size  256 | time 0.4223 ms | bandwidth 1430.12 GB/s
block_size  512 | time 0.4187 ms | bandwidth 1442.61 GB/s
block_size 1024 | time 0.4234 ms | bandwidth 1426.65 GB/s

BACKWARD PASS - AFTER
block_size   32 | time 0.3722 ms | bandwidth 1622.76 GB/s
block_size   64 | time 0.3290 ms | bandwidth 1835.54 GB/s
block_size  128 | time 0.3291 ms | bandwidth 1835.42 GB/s
block_size  256 | time 0.3290 ms | bandwidth 1835.81 GB/s
block_size  512 | time 0.3290 ms | bandwidth 1836.01 GB/s
block_size 1024 | time 0.3297 ms | bandwidth 1832.09 GB/s
```